# swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Marvel Characters REST Api
  description: API which returns i) all the Marvel character ids and ii) the id, name and description of a character.
host: localhost:8080
basePath: /
schemes:
  - http

definitions:
  CharacterIds:
    type: array
    items:
      type: integer
    example: [1011334, 1017100, 1009144, 1010699]
  Character:
    type: object
    required:
      - id
      - name
      - description
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
    example:
      id: 1009146
      name: "Abomination (Emil Blonsky)"
      description: "Formerly known as Emil Blonsky, a spy of Soviet Yugoslavian origin working for the KGB, the Abomination gained his powers after receiving a dose of gamma radiation similar to that which transformed Bruce Banner into the incredible Hulk."
  404NotFound:
    type: object
    required:
      - code
      - status
    properties:
      code:
        type: integer
      status:
        type: string
    example:
      code: 404
      status: "We couldn't find that character"

paths:
  /characters:
    get:
      description: Returns all Marvel character ids.
      responses:
        200:
          description: Successfully returned all Marvel character ids.
          schema:
            $ref: '#/definitions/CharacterIds'

  /characters/{characterId}:
    get:
      description: Returns id, name and description of a character.
      parameters:
        - name: characterId
          description: The character if of the character to get. If not specified, return all character ids.
          in: path
          type: integer
          required: true
      responses:
        200:
          description: Successfully returned the id, name and description of the character.
          schema:
            $ref: '#/definitions/Character'
        404:
          description: Invalid character id.
          schema:
            $ref: '#/definitions/404NotFound'